# backend/app/api/scenario.py
from datetime import date

from fastapi import APIRouter, Depends, Query

from ..exceptions import PermissionDeniedError
from ..schemas.scenario import (
    ScenarioComparisonRequest,
    ScenarioComparisonResponse,
    ScenarioDay,
    ScenarioListResponse,
    ScenarioResetResponse,
    ScenarioWrite,
)
from ..services.scenario import ScenarioService
from ..services.prediction import PredictionService
from .dependencies import (
    get_demo_persona,
    get_prediction_service,
    get_scenario_service,
    require_roles,
)


router = APIRouter(
    prefix="/api/demo/scenarios",
    tags=["Demo 场景"],
    dependencies=[Depends(require_roles("operator", "commuter"))],
)


def require_operator(persona: dict) -> None:
    if persona["role"] != "operator":
        raise PermissionDeniedError("仅 Demo 运营人员可修改未来场景")


@router.get("", response_model=ScenarioListResponse, summary="获取未来场景日历")
def list_scenarios(start: date | None = None, days: int = Query(14, ge=1, le=14), service: ScenarioService = Depends(get_scenario_service)) -> dict:
    return service.list(start, days)


@router.post("/compare", response_model=ScenarioComparisonResponse, summary="对比默认与候选场景 AI 方案")
def compare_scenarios(
    data: ScenarioComparisonRequest,
    service: ScenarioService = Depends(get_scenario_service),
    prediction_service: PredictionService = Depends(get_prediction_service),
) -> dict:
    return service.compare(data, prediction_service)


@router.post("/reset", response_model=ScenarioResetResponse, summary="恢复未来14天默认场景")
def reset_scenarios(service: ScenarioService = Depends(get_scenario_service), persona: dict = Depends(get_demo_persona)) -> dict:
    require_operator(persona)
    return service.reset()


@router.put("/{scenario_date}", response_model=ScenarioDay, summary="保存未来场景")
def save_scenario(scenario_date: date, data: ScenarioWrite, service: ScenarioService = Depends(get_scenario_service), persona: dict = Depends(get_demo_persona)) -> dict:
    require_operator(persona)
    return service.save(scenario_date, data)


@router.delete("/{scenario_date}", response_model=ScenarioDay, summary="恢复单日默认场景")
def delete_scenario(scenario_date: date, service: ScenarioService = Depends(get_scenario_service), persona: dict = Depends(get_demo_persona)) -> dict:
    require_operator(persona)
    return service.delete(scenario_date)