# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CrossBorder AI Demo API",
    "version": "1.3.0"
  },
  "paths": {
...
    "/api/demo/scenarios": {
      "get": {
        "tags": [
          "Demo 场景"
        ],
        "summary": "获取未来场景日历",
        "operationId": "list_scenarios_api_demo_scenarios_get",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 14,
              "minimum": 1,
              "default": 14,
              "title": "Days"
            }
          },
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioListResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/scenarios/compare": {
      "post": {
        "tags": [
          "Demo 场景"
        ],
        "summary": "对比默认与候选场景 AI 方案",
        "operationId": "compare_scenarios_api_demo_scenarios_compare_post",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioComparisonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioComparisonResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/scenarios/reset": {
      "post": {
        "tags": [
          "Demo 场景"
        ],
        "summary": "恢复未来14天默认场景",
        "operationId": "reset_scenarios_api_demo_scenarios_reset_post",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResetResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/scenarios/{scenario_date}": {
      "put": {
        "tags": [
          "Demo 场景"
        ],
        "summary": "保存未来场景",
        "operationId": "save_scenario_api_demo_scenarios__scenario_date__put",
        "parameters": [
          {
            "name": "scenario_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Scenario Date"
            }
          },
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioWrite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDay"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Demo 场景"
        ],
        "summary": "恢复单日默认场景",
        "operationId": "delete_scenario_api_demo_scenarios__scenario_date__delete",
        "parameters": [
          {
            "name": "scenario_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Scenario Date"
            }
          },
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDay"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/locations": {
      "get": {
        "tags": [
          "路线预测"
        ],
        "summary": "获取支持的出发地和目的地",
        "operationId": "locations_api_locations_get",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationsResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/predict": {
      "post": {
        "tags": [
          "路线预测"
        ],
        "summary": "比较四个口岸的路线方案",
        "description": "使用官方客流驱动的 AI V2.1、15分钟拥堵等级校准和确定性交通矩阵；模型或网络不可用时自动降级。",
        "operationId": "predict_api_predict_post",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "预测成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "user_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Action"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "details",
          "request_id",
          "category",
          "retryable"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "LocationOption": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "city"
        ],
        "title": "LocationOption"
      },
      "LocationsResponse": {
        "properties": {
          "origins": {
            "items": {
              "$ref": "#/components/schemas/LocationOption"
            },
            "type": "array",
            "title": "Origins"
          },
          "destinations": {
            "items": {
              "$ref": "#/components/schemas/LocationOption"
            },
            "type": "array",
            "title": "Destinations"
          },
          "directions": {
            "items": {
              "$ref": "#/components/schemas/DirectionOption"
            },
            "type": "array",
            "title": "Directions"
          }
        },
        "type": "object",
        "required": [
          "origins",
          "destinations",
          "directions"
        ],
        "title": "LocationsResponse"
      },
...
      "PredictionPreferences": {
        "properties": {
          "priority": {
            "$ref": "#/components/schemas/Priority",
            "default": "balanced"
          },
          "max_budget": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Budget"
          }
        },
        "type": "object",
        "title": "PredictionPreferences"
      },
...
      "PredictionRequest": {
        "properties": {
          "origin_id": {
            "type": "string",
            "minLength": 1,
            "title": "Origin Id"
          },
          "destination_id": {
            "type": "string",
            "minLength": 1,
            "title": "Destination Id"
          },
          "target_time": {
            "type": "string",
            "format": "date-time",
            "title": "Target Time"
          },
          "preferences": {
            "$ref": "#/components/schemas/PredictionPreferences"
          }
        },
        "type": "object",
        "required": [
          "origin_id",
          "destination_id",
          "target_time"
        ],
        "title": "PredictionRequest"
      },
      "PredictionResponse": {
        "properties": {
          "query": {
            "$ref": "#/components/schemas/PredictionQuery"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/PortPrediction"
            },
            "type": "array",
            "title": "Ports"
          },
          "recommended": {
            "type": "string",
            "title": "Recommended"
          },
          "recommended_port_id": {
            "type": "string",
            "title": "Recommended Port Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "confidence_level": {
            "type": "number",
            "title": "Confidence Level"
          },
          "demo_notice": {
            "type": "string",
            "title": "Demo Notice"
          },
          "data_sources": {
            "items": {
              "$ref": "#/components/schemas/DataSourceStatus"
            },
            "type": "array",
            "title": "Data Sources"
          },
          "data_version": {
            "type": "string",
            "title": "Data Version"
          },
          "direction": {
            "$ref": "#/components/schemas/TravelDirection"
          },
          "forecast_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Run Id"
          },
          "prediction_engine": {
            "type": "string",
            "title": "Prediction Engine"
          },
          "scenario": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scenario"
          }
        },
        "type": "object",
        "required": [
          "query",
          "ports",
          "recommended",
          "recommended_port_id",
          "reason",
          "warnings",
          "generated_at",
          "model_version",
          "confidence_level",
          "demo_notice",
          "data_sources",
          "data_version",
          "direction",
          "prediction_engine",
          "scenario"
        ],
        "title": "PredictionResponse"
      },
...
      "ScenarioComparisonPort": {
        "properties": {
          "port_id": {
            "type": "string",
            "title": "Port Id"
          },
          "port_name": {
            "type": "string",
            "title": "Port Name"
          },
          "baseline_wait_minutes": {
            "type": "integer",
            "title": "Baseline Wait Minutes"
          },
          "candidate_wait_minutes": {
            "type": "integer",
            "title": "Candidate Wait Minutes"
          },
          "wait_delta_minutes": {
            "type": "integer",
            "title": "Wait Delta Minutes"
          },
          "baseline_late_risk_percent": {
            "type": "integer",
            "title": "Baseline Late Risk Percent"
          },
          "candidate_late_risk_percent": {
            "type": "integer",
            "title": "Candidate Late Risk Percent"
          },
          "late_risk_delta_percent": {
            "type": "integer",
            "title": "Late Risk Delta Percent"
          },
          "total_time_delta_minutes": {
            "type": "integer",
            "title": "Total Time Delta Minutes"
          }
        },
        "type": "object",
        "required": [
          "port_id",
          "port_name",
          "baseline_wait_minutes",
          "candidate_wait_minutes",
          "wait_delta_minutes",
          "baseline_late_risk_percent",
          "candidate_late_risk_percent",
          "late_risk_delta_percent",
          "total_time_delta_minutes"
        ],
        "title": "ScenarioComparisonPort"
      },
      "ScenarioComparisonRequest": {
        "properties": {
          "origin_id": {
            "type": "string",
            "minLength": 1,
            "title": "Origin Id"
          },
          "destination_id": {
            "type": "string",
            "minLength": 1,
            "title": "Destination Id"
          },
          "target_time": {
            "type": "string",
            "format": "date-time",
            "title": "Target Time"
          },
          "preferences": {
            "$ref": "#/components/schemas/PredictionPreferences"
          },
          "scenario": {
            "$ref": "#/components/schemas/ScenarioWrite"
          }
        },
        "type": "object",
        "required": [
          "origin_id",
          "destination_id",
          "target_time",
          "scenario"
        ],
        "title": "ScenarioComparisonRequest"
      },
      "ScenarioComparisonResponse": {
        "properties": {
          "baseline": {
            "$ref": "#/components/schemas/PredictionResponse"
          },
          "candidate": {
            "$ref": "#/components/schemas/PredictionResponse"
          },
          "recommended_changed": {
            "type": "boolean",
            "title": "Recommended Changed"
          },
          "baseline_recommended_port_id": {
            "type": "string",
            "title": "Baseline Recommended Port Id"
          },
          "candidate_recommended_port_id": {
            "type": "string",
            "title": "Candidate Recommended Port Id"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/ScenarioComparisonPort"
            },
            "type": "array",
            "title": "Ports"
          }
        },
        "type": "object",
        "required": [
          "baseline",
          "candidate",
          "recommended_changed",
          "baseline_recommended_port_id",
          "candidate_recommended_port_id",
          "ports"
        ],
        "title": "ScenarioComparisonResponse"
      },
      "ScenarioDay": {
        "properties": {
          "weather": {
            "$ref": "#/components/schemas/ScenarioWeather"
          },
          "is_holiday": {
            "type": "boolean",
            "title": "Is Holiday",
            "default": false
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ScenarioEvent"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Events"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "is_override": {
            "type": "boolean",
            "title": "Is Override"
          }
        },
        "type": "object",
        "required": [
          "weather",
          "date",
          "version",
          "is_override"
        ],
        "title": "ScenarioDay"
      },
      "ScenarioEvent": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Name"
          },
          "preset": {
            "type": "string",
            "maxLength": 30,
            "title": "Preset",
            "default": "custom"
          },
          "direction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TravelDirection"
              },
              {
                "type": "null"
              }
            ]
          },
          "affected_ports": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Affected Ports"
          },
          "start_time": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "End Time"
          },
          "impact": {
            "$ref": "#/components/schemas/ScenarioImpact"
          }
        },
        "type": "object",
        "required": [
          "name",
          "affected_ports",
          "start_time",
          "end_time",
          "impact"
        ],
        "title": "ScenarioEvent"
      },
...
      "ScenarioListResponse": {
        "properties": {
          "start": {
            "type": "string",
            "format": "date",
            "title": "Start"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioDay"
            },
            "type": "array",
            "title": "Scenarios"
          },
          "weather_options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weather Options"
          },
          "event_presets": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Event Presets"
          }
        },
        "type": "object",
        "required": [
          "start",
          "days",
          "scenarios",
          "weather_options",
          "event_presets"
        ],
        "title": "ScenarioListResponse"
      },
      "ScenarioResetResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioDay"
            },
            "type": "array",
            "title": "Scenarios"
          }
        },
        "type": "object",
        "required": [
          "success",
          "scenarios"
        ],
        "title": "ScenarioResetResponse"
      },
      "ScenarioWeather": {
        "type": "string",
        "enum": [
          "clear",
          "rain",
          "heavy_rain",
          "thunderstorm"
        ],
        "title": "ScenarioWeather"
      },
      "ScenarioWrite": {
        "properties": {
          "weather": {
            "$ref": "#/components/schemas/ScenarioWeather"
          },
          "is_holiday": {
            "type": "boolean",
            "title": "Is Holiday",
            "default": false
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ScenarioEvent"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "weather"
        ],
        "title": "ScenarioWrite"
      },
...
    }
  }
}